requirements.txt
fastapi
pydantic-settings
SQLAlchemy
email_validator
uvicorn
rich
asgi-correlation-id
python-json-logger
python-jose
bcrypt
aiosqlite
asyncpg
python-multipart
greenlet
pytest
httpx
Jinja2
pytailwindcss
python-dateutil
psycopg2-binary
pytz
overrides